refactor(extensions): split sandbox/runtime helper modules
